logger: restore Logger and logging callback on integ. test teardown
When integration test gets constructed, it can set global `cb_` and
`data_` to temporaries. After the test finishes and temporaries are
destroyed, the next tests ctor may try to refer to them (before it
sets the globals by itself).

This happened in the ctor of `Random`, which logged a message in
case of failure. Sometimes this manifested as segfault, sometimes
as a deadlock.

This fix restores the old logger values on test teardown - most
likely to stderr.

Fixes #13